Refactor: split GameSessionManager into Persistence/Creation managers. Shared package for (most) Game*Manager.
GamePersistenceManager now handles persistence of saved games, while GameCreationManager creates Game and related objects from persisted data or GameSetups.
