+++ app/src/main/java/com/peaceray/codeword/domain/manager/game/setup/impl/setup/versioned/language/CodeLanguageDetailsFactory.kt
@@ -1,9 +1,9 @@
-package com.peaceray.codeword.domain.manager.game.impl.setup.versioned.language
+package com.peaceray.codeword.domain.manager.game.setup.impl.setup.versioned.language
 
 import com.peaceray.codeword.data.model.code.CodeLanguage
 import com.peaceray.codeword.data.model.code.CodeLanguageDetails
-import com.peaceray.codeword.domain.manager.game.impl.setup.versioned.seed.SeedVersion
-import com.peaceray.codeword.domain.manager.game.impl.setup.versioned.seed.VersionedBySeed
+import com.peaceray.codeword.domain.manager.game.setup.impl.setup.versioned.seed.SeedVersion
+import com.peaceray.codeword.domain.manager.game.setup.impl.setup.versioned.seed.VersionedBySeed
 import java.util.SortedSet
 
 /**
